% Auto-generated: do not edit by hand
\name{dashTabulator}

\alias{dashTabulator}

\title{DashTabulator component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dashTabulator(id=NULL, columns=NULL, data=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{columns}{Unnamed list. A label that will be printed when this component is rendered.}

\item{data}{Unnamed list. The value displayed in the input.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

